specification based changes synced with latest swagger updates

+++ lib/sib-api-v3-sdk/models/get_account_plan.rb
@@ -60,7 +60,7 @@ module SibApiV3Sdk
       {
         :'type' => :'String',
         :'credits_type' => :'String',
-        :'credits' => :'Integer'
+        :'credits' => :'Float'
       }
     end
